+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v 12 "" (WIP)
 - fix: Open Downloads after download complete
+- fix: Custom search engine UI/UX #1113
 
 
 ### v 11 "Bertha von Suttner"